% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egal.R
\name{al_egal}
\alias{al_egal}
\title{Implementation of Exploration Guided Active Learning (EGAL)}
\usage{
al_egal(
  samples_tb,
  sim_method = "correlation",
  alpha = NULL,
  beta = alpha,
  w = 0.5
)
}
\arguments{
\item{samples_tb}{A sits tibble with both labelled and unlabelled
samples (i.e. NA).}

\item{sim_method}{A character. A method for computing the similarity
among samples. See proxy::simil for details.}

\item{alpha}{A double. It controls the radius of the neighborhood
used in the estimation of sample density.}

\item{beta}{A double. It controls the radius of the neighborhood
used in the estimation of the sample candidate set. A
bigger beta gives a bigger set. By default is set to
be equal to alpha. If NULL, it is computed internally.}

\item{w}{A numeric (between 0 and 1) only used when beta is
NULL. This proportion parameter balances the influence
of diversity and density in the selection strategy.
When w is 0, EGAL becomes a pure-diversity and when w
is 1, EGAL becomes a pure density-based sampling
algorithm.}
}
\value{
\preformatted{           A sits tibble with the EGAL metric. This metric
                   ranks samples based on their density and diversity.
                   Those samples with highest EGAL should be selected
                   first for labeling.
}
}
\description{
Active Learning improves the results of a classification by
feeding the classifier with informative samples. The Exploration Guided
Active Learning (EGAL) ranks a set of samples based on their density and
diversity; those samples with a larger EGAL metric should be submitted first
to a human expert (the oracle) for classification.

This function receives a sits tibble with time series samples and it computes
the EGAL metric on it. However, this function doesn't guarantee the order of
the returned samples.
}
\references{
Hu, R., Jane Delany, S., & Mac Namee, B. (2010). EGAL:
Exploration Guided Active Learning for TCBR. In Lecture Notes in Computer
Science (including subseries Lecture Notes in Artificial Intelligence and
Lecture Notes in Bioinformatics): Vol. 6176 LNAI (pp. 156–170). doi:
10.1007/978-3-642-14274-1_13
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
