% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{al_random_sampling}
\alias{al_random_sampling}
\title{Implementation of Active Learning using a random sampling strategy.}
\usage{
al_random_sampling(samples_tb, sits_method, multicores = 1)
}
\arguments{
\item{samples_tb}{A sits tibble.}

\item{sits_method}{A sits model specification.}

\item{multicores}{The number of cores available for active learning.}
}
\value{
\preformatted{       A sits tibble with metrics. Entropy is a measure of the
               amount of information in the probabilities of each label;
               the samples with largest entropy are the best candidates
               for labeling by human experts. Least Confidence is the
               difference between the most confident prediction and 100\%
               confidence normalized by the number of labels. Margin of
               Confidence is the difference between the two most
               confident predictions. Ratio of Confidence is the ratio
               between the top two most confident predictions.
}
}
\description{
Active Learning improves the results of a classification by
feeding the classifier with informative samples. This function receives a
set of labelled and unlabelled samples. The labelled samples are used to
train a model which is then used to classify the unlabelled samples and then
it computes some metrics on the results. These metrics are useful for
selecting the samples to be sent to an expert (oracle) for labeling.

This function receives a sits tibble and it returns it with metrics.
However, this function doesn't guarantee the order of the returned samples.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
