% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{al_s2}
\alias{al_s2}
\title{Implementation of Shortest Shortest Path (S2) algorithm}
\usage{
al_s2(
  samples_tb,
  sim_method = "correlation",
  closest_n = 6,
  mode = "active_learning"
)
}
\arguments{
\item{samples_tb}{A sits tibble with both labelled and unlabelled
samples (i.e. NA).}

\item{sim_method}{A character. A method for computing the similarity
among samples. See proxy::simil for details.}

\item{closest_n}{An integer. The number of most similar samples to
keep while building a similarity graph of the
samples.}

\item{mode}{A character telling if the functin runs on either the
"active_learning" or "semi_supervised_learning" mode.}
}
\value{
\preformatted{           A sits tibble with either the samples to be sent to
                   the oracle (mode "semi_supervised_learning", column
}
}
\description{
Semi-Supervised Learning is a set of classification methods
that use both labelled and unlabelled samples. The Shortest Shortest Path
(S2) algorithm uses a undirected graph of the samples and iteratively
removes edges, trying to identify the boundaries of each class in the graph.

This function receives a sits tibble with time series samples and it returns
a sits tibble with either the unlabelled samples to be sent to the oracle
or the label of each sample.
}
\references{
Dasarathy, G., Nowak, R., & Zhu, X. (2015). S2: An efficient
graph based active learning algorithm with application to nonparametric
classification. Journal of Machine Learning Research, 40(2015), 1–20.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
