% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.al_check_time_series}
\alias{.al_check_time_series}
\title{Check the time series in a sits tibble.}
\usage{
.al_check_time_series(sits_tb)
}
\arguments{
\item{s_labelled_tb}{A sits tibble.}
}
\value{
\preformatted{           TRUE or it throws an error.
}
}
\description{
This function returns TRUE of the time series in the given sits
tibble are valid; otherwise it throws an error.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
