% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.al_count_na}
\alias{.al_count_na}
\title{Count the number of NAs in the time series of a sits tibble.}
\usage{
.al_count_na(sits_tb)
}
\arguments{
\item{sits_tb}{A sits tibble.}
}
\value{
\preformatted{   An integer.
}
}
\description{
Count the number of NAs in the time series of each observation
in a sits tibble.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
