% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egal.R
\name{.al_egal}
\alias{.al_egal}
\title{Implementation of Exploration Guided Active Learning (EGAL)}
\usage{
.al_egal(
  s_labelled_tb,
  s_unlabelled_tb,
  sim_method = "correlation",
  alpha = NULL,
  beta = alpha,
  w = 0.5
)
}
\arguments{
\item{s_labelled_tb}{A sits tibble with labelled samples.}

\item{s_unlabelled_tb}{A sits tibble with unlabelled samples.}

\item{sim_method}{A character. A method for computing the similarity
among samples as described in the package proxy. If
the method is registered in proxy::pr_DB as a
distance, then the inverse of the distance is used
(e.g. 1/distance).
proxy::pr_DB as a distance, then the}

\item{alpha}{A double. It controls the radius of the neighborhood
used in the estimation of sample density.}

\item{beta}{A double. It controls the radius of the neighborhood
used in the estimation of the sample candidate set. A
bigger beta gives a bigger set. By default is set to
be equal to alpha. If NULL, it is computed internally.}

\item{w}{A numeric (between 0 and 1) only used when beta is
NULL. This proportion parameter balances the influence
of diversity and density in the selection strategy.
When w is 0, EGAL becomes a pure-diversity
and when w is 1, EGAL becomes a pure density-based
sampling algorithm.}
}
\value{
\preformatted{           A sits tibble with the EGAL metric. This metric
                   ranks samples based on their density and diversity.
                   Those samples with highest EGAL should be selected
                   first for labeling.
}
}
\description{
This function returns a sits tibble with their score in the
metric Exploration Guided Active Learning (EGAL). Samples with a larger EGAL
metric should be submitted first to a human expert for classification.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
