% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egal.R
\name{.al_egal_update_beta}
\alias{.al_egal_update_beta}
\title{Compute a new value for the beta parameter.}
\usage{
.al_egal_update_beta(S_mat, w)
}
\arguments{
\item{S_mat}{A numeric matrix representing the similarity between unlabelled
(rows) and labelled (columns) samples.}

\item{w}{A numeric (between 0 and 1). A proportion parameter.}
}
\value{
\preformatted{ A length-one numeric.
}
}
\description{
Compute a new beta parameter for the Exploration Guided Active
Learning (EGAL) algorithm.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
