% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{.al_get_random_points}
\alias{.al_get_random_points}
\alias{al_get_random_points}
\title{Get a sits tibble of random points in the given data cube.}
\usage{
al_get_random_points(data_cube, n_samples, multicores)
}
\arguments{
\item{data_cube}{A sits data cube.}

\item{n_samples}{The number of random points to take. The returned
sample size may be less than the requested size.}

\item{multicores}{The number of cores available for active learning.}
}
\value{
\preformatted{      A sits tibble of samples, including their time series.
}
}
\description{
Get a sits tibble of random points in the given data cube.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
