% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{.al_rs}
\alias{.al_rs}
\title{Implementation of Active Learning using random sampling}
\usage{
.al_rs(s_labelled_tb, s_unlabelled_tb, sits_method, multicores)
}
\arguments{
\item{s_labelled_tb}{A sits tibble with labelled samples.}

\item{s_unlabelled_tb}{A sits tibble with unlabelled samples.}

\item{sits_method}{A sits model specification.}

\item{multicores}{The number of cores available for active learning.}
}
\value{
\preformatted{       A sits tibble with metrics. Entropy is a measure of the
               amount of information in the probabilities of each label;
               the samples with largest entropy are the best candidates
               for labeling by human experts. Least Confidence is the
               difference between the most confident prediction and 100\%
               confidence normalized by the number of labels. Margin of
               Confidence is the difference between the two most
               confident predictions. Ratio of Confidence is the ratio
               between the top two most confident predictions.
}
}
\description{
This function returns a sits tibble with metrics.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
