% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{.al_s2}
\alias{.al_s2}
\title{Implementation of Shortest Shortest Path (S2) algorithm}
\usage{
.al_s2(s_labelled_tb, s_unlabelled_tb, sim_method, closest_n, mode)
}
\arguments{
\item{s_labelled_tb}{A sits tibble with labelled samples.}

\item{s_unlabelled_tb}{A sits tibble with unlabelled samples.}

\item{sim_method}{A character. A method for computing the similarity
among samples. See proxy::simil for details.}

\item{closest_n}{An integer. The number of most similar samples to
keep while building the graph.}

\item{mode}{A character telling if the functin runs on either the
"active_learning" or "semi_supervised_learning" mode.}
}
\value{
\preformatted{           A sits tibble with either the samples to be sent to
                   the oracle (mode "semi_supervised_learning", column
}
}
\description{
The S2 algorithm uses a undirected graph of the samples and
iteratively removes edges, trying to identify the boundaries of each class
in the graph.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
