% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{.al_s2_bild_closest_vertex_graph}
\alias{.al_s2_bild_closest_vertex_graph}
\title{Build a matrix representing a graph made of the closest verteces.}
\usage{
.al_s2_bild_closest_vertex_graph(samples_tb, sim_method, closest_n)
}
\arguments{
\item{samples_tb}{A sits tibble with both labelled and unlabelled
samples (i.e. NA).}

\item{sim_method}{A character. A method for computing the similarity
among samples. See proxy::simil for details.}

\item{closest_n}{An integer. The number of most similar samples to
link in the graph.}
}
\value{
\preformatted{           A lower triangular matrix. The matrix's upper
                   triangular part along its principal diagonal are set
                   to NA. The rows in the matrix's lower triangular part
                   are non-NA for the closest_n elements.
}
}
\description{
This function takes a sits tibble and returns a matrix
representation of a similarity graph where only the most similar samples
are connected.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
