% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{.al_s2_mssp}
\alias{.al_s2_mssp}
\title{Compute the MSSP subroutine.}
\usage{
.al_s2_mssp(G, L)
}
\arguments{
\item{G}{A igraph object.}

\item{L}{A list of lists containing the ids of the vertexes in G_mt and
their labels.}
}
\value{
\preformatted{A vector. The vertex ids of the mid-points along the shortest
        paths between the elements of L.
}
}
\description{
This function takes a graph and a set of labelled vertexes and
computes the length of the shortest paths and the points along them.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
