% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2.R
\name{.al_s2_remove_mismatch_edges}
\alias{.al_s2_remove_mismatch_edges}
\title{Remove mismatching edges from a graph.}
\usage{
.al_s2_remove_mismatch_edges(G)
}
\arguments{
\item{G}{An igraph object whose veteces have a label attribute.}
}
\value{
\preformatted{   A graph with the same or less edges than G.
}
}
\description{
This function takes a graph and removes the edges with
different labels at their extremes.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
