% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{.sits_compute_metrics}
\alias{.sits_compute_metrics}
\title{Compute metrics of active learning using random sampling.}
\usage{
.sits_compute_metrics(x, points_tb)
}
\arguments{
\item{x}{A length-one integer.}

\item{points_tb}{A sits tibble.}
}
\value{
\preformatted{      A sits tibble of samples, including their time series.
}
}
\description{
Get a sits tibble of random points in the given data cube.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
\keyword{internal}
