% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/query-committee.r
\name{query_committee}
\alias{query_committee}
\title{Active learning with "Query by Committee"}
\usage{
query_committee(x, y, committee, disagreement = c("kullback", "vote_entropy",
  "post_entropy"), num_query = 1, num_cores = 1, ...)
}
\arguments{
\item{x}{a matrix containing the labeled and unlabeled data}

\item{y}{a vector of the labels for each observation in \code{x}. Use
\code{NA} for unlabeled.}

\item{committee}{a list containing the committee of classifiers. See details
for the required format.}

\item{disagreement}{a string that contains the disagreement measure among the
committee members. See above for details.}

\item{num_query}{the number of observations to be queried.}
}
\value{
a list that contains the least_certain observation and miscellaneous
results. See above for details.
}
\description{
The 'query by committee' approach to active learning uitilizes a committee of
\code{C} classifiers that are each trained on the labeled training data. Our
goal is to "query the oracle" with the observations that have the maximum
disagreement among the \code{C} trained classifiers.
}
\details{
Note that this approach is similar to "Query by Bagging" (QBB), but each
committee member is specified by the user. With the QBB approach, only one
supervised classifier is specified by the user, and each committee member is
trained on a resampled subset of  the labeled training data. Also, note that
we have implemented QBB as \code{\link{query_by_bagging}}.

To determine maximum disagreement among committee committe members, we have
implemented three approaches:

\describe{
\item{kullback}{query the unlabeled observation that maximizes the
Kullback-Leibler divergence between the label distributions of any one
committe member and the consensus}
\item{vote_entropy}{query the unlabeled observation that maximizes the vote
entropy among all commitee members}
\item{post_entropy}{query the unlabeled observation that maximizes the entropy
of average posterior probabilities of all committee members}
}

The \code{disagreement} argument must be one of the three: \code{kullback} is
the default.

To calculate the committee disagreement, we use the formulae from Dr. Burr
Settles' "Active Learning Literature Survey" available on his website.
At the time this function was coded, the literature survey had last been
updated on January 26, 2010.

To specify the committee members, we require a character vector of classifiers
in the \code{committee} argument with elements corresponding to each
supervised classifier (each committee member). The classifiers must match the
naming in the \code{caret} package.

Unlabeled observations in \code{y} are assumed to have \code{NA} for a label.

It is often convenient to query unlabeled observations in batch. By default,
we query the unlabeled observation with the largest disagreement measure
value. With the \code{num_query} the user can specify the number of
observations to return in batch. If there are ties in the disagreement measure
values, they are broken by the order in which the unlabeled observations are
given.
}
\examples{
x <- iris[, -5]
y <- iris[, 5]

# For demonstration, suppose that few observations are labeled in 'y'.
y <- replace(y, -c(1:12, 51:62, 101:112), NA)

committee <- c("lda", "qda", "rda")
query_committee(x=x, y=y, committee=committee, num_query=3)$query
query_committee(x=x, y=y, committee=committee, disagreement="post_entropy",
                num_query=5)$query
}

