% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query-random.r
\name{query_random}
\alias{query_random}
\title{Active Learning with Random Querying}
\usage{
query_random(y, num_query = 1)
}
\arguments{
\item{y}{a vector of the classification labels for each observation in a data
set. Use \code{NA} for unlabeled observations.}

\item{num_query}{the number of observations to be be queried.}
}
\value{
a list indicates which observations to \code{query}
}
\description{
A simple, naive approach is to query unlabeled observations chosen at random.
Random querying is often a baseline to which proposed active learning methods
are compared.
}
\details{
Unlabeled observations in \code{y} are assumed to have \code{NA} for a label.
It is often convenient to query unlabeled observations in batch. With the
\code{num_query} the user can specify the number of observations to return in
batch.
}

