% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_sampling.R
\name{sits_al_random_sampling}
\alias{sits_al_random_sampling}
\title{Implementation of Active Learning using a random sampling stategy.}
\usage{
sits_al_random_sampling(
  samples_tb,
  sits_method,
  data_cube,
  n_samples = 1000,
  multicores = 2
)
}
\arguments{
\item{samples_tb}{A sits tibble.}

\item{sits_method}{A sits model specification.}

\item{data_cube}{A sits data cube.}

\item{n_samples}{The number of random points to take.}

\item{multicores}{The number of cores available for active learning.}
}
\value{
\preformatted{       A sits tibble with metrics. Entropy is a measure of the
               amount of information in the probabilities of each label;
               the samples with largest entropy are the best candidates
               for labeling by human experts. Least Confidence is the
               difference between the most confident prediction and 100\%
               confidence normalized by the number of labels. Margin of
               Confidence is the difference between the two most
               confident predictions. Ratio of Confidence is the ratio
               between the top two most confident predictions.
}
}
\description{
Active Learning improves the results of a classification by
feeding the classifier with informative samples. This function returns a sits
tibble of new samples selected at random and some metrics useful for
selecting the samples to be sent to an expert for labeling.

These new_samples could be merged to the original samples (samples_tb) to
increase the number of training data set.
}
\author{
Alber Sanchez, \email{alber.ipia@inpe.br}
}
