% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disagreement.r
\name{vote_entropy}
\alias{vote_entropy}
\title{Computes entropy of committee's classifications}
\usage{
vote_entropy(x, type = "class", entropy_method = "ML")
}
\description{
Computes the disagreement measure for each of the unlabeled observations
based on the either the predicted class labels or the posterior
probabilities of class membership.
}

